requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
sqlparse==0.4.3
djangorestframework==3.14.0
psycopg2==2.9.5
